fix(xtask): address pull-subrepo conflict-handling review findings
